stage

+++ player/player.h
@@ -3,15 +3,16 @@
 
 # include "../libft/libft.h"
 # include "../map/map.h"
+# include "../constants.h"
 
 typedef struct s_player
 {
-	float	x;
-	float	y;
-	float	a;
+	double	x;
+	double	y;
+	double	a;
 }	t_player;
 
 t_player	*new_player(t_map *map);
-void	print_player(t_player *pl);
+void		print_player(t_player *pl);
 
 #endif
